requirements.txt
ib_insync==0.9.86
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
python-dotenv==1.0.1
APScheduler==3.10.4
python-telegram-bot==21.3
sqlalchemy==2.0.30
requests==2.32.3
